requirements.txt
streamlit>=1.31.0
streamlit-folium>=0.17.0
folium>=0.15.1
jaxa>=0.1.0
Pillow>=10.0.0
matplotlib>=3.8.0
